feat(core): move runtime-to-client bridge
