+++ packages/core/src/store/clients/suggestions.ts
@@ -1,6 +1,6 @@
 import { resource, tapState, withKey } from "@assistant-ui/tap";
 import type { ClientOutput } from "../types/client";
-import { tapClientLookup } from "../tap-client-lookup";
+import { tapClientLookup } from "../utils/tap-client-lookup";
 import type { SuggestionsState, SuggestionState } from "../scopes";
 
 export type SuggestionConfig =
