+++ packages/core/src/store/runtime-clients/attachment-runtime-client.ts
@@ -1,7 +1,7 @@
 import { resource } from "@assistant-ui/tap";
 import { type ClientOutput } from "../";
 import { AttachmentRuntime } from "../../runtime";
-import { tapSubscribable } from "../utils/tap-subscribable";
+import { tapSubscribable } from "./tap-subscribable";
 
 export const AttachmentRuntimeClient = resource(
   ({ runtime }: { runtime: AttachmentRuntime }): ClientOutput<"attachment"> => {
