+++ packages/core/src/store/runtime-clients/message-part-runtime-client.ts
@@ -1,7 +1,7 @@
 import { resource } from "@assistant-ui/tap";
 import { type ClientOutput } from "../";
 import { MessagePartRuntime } from "../../runtime";
-import { tapSubscribable } from "../utils/tap-subscribable";
+import { tapSubscribable } from "./tap-subscribable";
 
 export const MessagePartClient = resource(
   ({ runtime }: { runtime: MessagePartRuntime }): ClientOutput<"part"> => {
