+++ packages/core/src/store/utils/tap-client-lookup.ts
@@ -4,7 +4,7 @@ import {
   tapResources,
   type ResourceElement,
 } from "@assistant-ui/tap";
-import type { ClientMethods } from "./types/client";
+import type { ClientMethods } from "../types/client";
 import { ClientResource } from "./tap-client-resource";
 import { wrapperResource } from "./wrapper-resource";
